Move the v3io_tsdb_reader module into v3io_connector
